[util] Add DefaultBottomSheet
