+++ ui-component/src/main/java/com/example/ui_component/DefaultBottomSheet.kt
@@ -1,0 +1,21 @@
+package com.example.ui_component
+
+import androidx.compose.material3.BottomSheetDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.rememberModalBottomSheetState
+import androidx.compose.runtime.Composable
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun DefaultBottomSheet(onDismiss: () -> Unit, content: @Composable () -> Unit) {
+
+    val modalBottomSheetState = rememberModalBottomSheetState()
+    ModalBottomSheet(
+        onDismissRequest = { onDismiss() },
+        sheetState = modalBottomSheetState,
+        dragHandle = { BottomSheetDefaults.DragHandle() },
+    ) {
+        content()
+    }
+}
